Remove unused headers,replaced stateChanged,fixed temporary.first() reference error
