+++ src/CMakeLists.txt
@@ -169,7 +169,7 @@ set(SOURCES
     widgets/SearchableDockWidget.cpp
     common/CutterSearchable.cpp
     widgets/SearchableTextEdit.cpp
-    dialogs/preferences/SymbolServers.cpp
+    dialogs/preferences/SymbolsOptionsWidget.cpp
 )
 set(HEADER_FILES
     core/Cutter.h
@@ -348,7 +348,7 @@ set(HEADER_FILES
     widgets/SearchableDockWidget.h
     common/CutterSearchable.h
     widgets/SearchableTextEdit.h
-    dialogs/preferences/SymbolServers.h
+    dialogs/preferences/SymbolsOptionsWidget.h
 )
 set(UI_FILES
     dialogs/AboutDialog.ui
@@ -430,7 +430,7 @@ set(UI_FILES
     dialogs/RegisterProfileDialog.ui
     dialogs/EditRegProfileDialog.ui
     widgets/SearchBarWidget.ui
-    dialogs/preferences/SymbolServers.ui
+    dialogs/preferences/SymbolsOptionsWidget.ui
 )
 set(QRC_FILES
     resources.qrc
