+++ src/dialogs/preferences/SymbolsOptionsWidget.ui
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>SymbolServers</class>
- <widget class="QWidget" name="SymbolServers">
+ <class>SymbolsOptionsWidget</class>
+ <widget class="QWidget" name="SymbolsOptionsWidget">
   <property name="geometry">
    <rect>
     <x>0</x>
