+++ src/resources.qrc
@@ -125,8 +125,8 @@
         <file>img/icons/reset.svg</file>
         <file>img/icons/down_light.svg</file>
         <file>img/icons/delete_light.svg</file>
-        <file>img/icons/symbol_server.svg</file>
-        <file>img/icons/light/symbol_server.svg</file>
+        <file>img/icons/symbol_options.svg</file>
+        <file>img/icons/light/symbol_options.svg</file>
         <file>img/icons/light/edit.svg</file>
         <file>img/icons/edit.svg</file>
     </qresource>
